+++ src/components/cards/measuring.tsx
@@ -17,7 +17,7 @@ import {
     triggerLocalRefresh,
 } from "@/lib/context";
 import { cn } from "@/lib/utils";
-import { iconColors } from "@/maps/api";
+import { ICON_COLORS } from "@/maps/api/index";
 import { determineMeasuringBoundary } from "@/maps/measuring";
 import {
     determineUnionizedStrings,
@@ -216,7 +216,7 @@ export const MeasuringQuestionComponent = ({
                     "text-2xl font-semibold font-poppins",
                 )}
                 style={{
-                    backgroundColor: iconColors[data.color],
+                    backgroundColor: ICON_COLORS[data.color],
                     color: data.color === "gold" ? "black" : undefined,
                 }}
             >
