+++ src/maps/api/cache.ts
@@ -1,0 +1,68 @@
+import _ from "lodash";
+import { toast } from "react-toastify";
+
+import { CacheType } from "./types";
+
+const determineQuestionCache = _.memoize(() => caches.open(CacheType.CACHE));
+const determineZoneCache = _.memoize(() => caches.open(CacheType.ZONE_CACHE));
+const determinePermanentCache = _.memoize(() =>
+    caches.open(CacheType.PERMANENT_CACHE),
+);
+
+export const determineCache = async (cacheType: CacheType) => {
+    switch (cacheType) {
+        case CacheType.CACHE:
+            return await determineQuestionCache();
+        case CacheType.ZONE_CACHE:
+            return await determineZoneCache();
+        case CacheType.PERMANENT_CACHE:
+            return await determinePermanentCache();
+    }
+};
+
+export const cacheFetch = async (
+    url: string,
+    loadingText?: string,
+    cacheType: CacheType = CacheType.CACHE,
+) => {
+    try {
+        const cache = await determineCache(cacheType);
+
+        const cachedResponse = await cache.match(url);
+        if (cachedResponse) return cachedResponse;
+
+        if (loadingText) {
+            return toast.promise(
+                async () => {
+                    const response = await fetch(url);
+                    await cache.put(url, response.clone());
+                    return response;
+                },
+                {
+                    pending: loadingText,
+                },
+            );
+        }
+
+        const response = await fetch(url);
+        await cache.put(url, response.clone());
+        return response;
+    } catch (e) {
+        console.log(e); // Probably a caches not supported error
+
+        return fetch(url);
+    }
+};
+
+export const clearCache = async (cacheType: CacheType = CacheType.CACHE) => {
+    try {
+        const cache = await determineCache(cacheType);
+        await cache.keys().then((keys) => {
+            keys.forEach((key) => {
+                cache.delete(key);
+            });
+        });
+    } catch (e) {
+        console.log(e); // Probably a caches not supported error
+    }
+};
