+++ src/maps/api/constants.ts
@@ -1,0 +1,30 @@
+import type { APILocations } from "@/maps/schema";
+
+export const OVERPASS_API = "https://overpass-api.de/api/interpreter";
+export const GEOCODER_API = "https://photon.komoot.io/api/";
+
+export const ICON_COLORS = {
+    black: "#3D3D3D",
+    blue: "#2A81CB",
+    gold: "#FFD326",
+    green: "#2AAD27",
+    grey: "#7B7B7B",
+    orange: "#CB8427",
+    red: "#CB2B3E",
+    violet: "#9C2BCB",
+};
+
+export const LOCATION_FIRST_TAG: {
+    [key in APILocations]: "amenity" | "tourism" | "leisure" | "diplomatic";
+} = {
+    aquarium: "tourism",
+    hospital: "amenity",
+    museum: "tourism",
+    theme_park: "tourism",
+    zoo: "tourism",
+    cinema: "amenity",
+    library: "amenity",
+    golf_course: "leisure",
+    consulate: "diplomatic",
+    park: "leisure",
+};
