+++ src/maps/api/overpass.ts
@@ -1,63 +1,22 @@
 import * as turf from "@turf/turf";
-import type { LatLngTuple } from "leaflet";
 import _ from "lodash";
 import osmtogeojson from "osmtogeojson";
-import { toast } from "react-toastify";
 
 import {
     additionalMapGeoLocations,
     mapGeoLocation,
     polyGeoJSON,
 } from "@/lib/context";
+
+import { cacheFetch } from "./cache";
+import { LOCATION_FIRST_TAG, OVERPASS_API } from "./constants";
 import type {
     EncompassingTentacleQuestionSchema,
-    Question,
-} from "@/maps/schema";
-import type {
-    APILocations,
     HomeGameMatchingQuestions,
     HomeGameMeasuringQuestions,
-} from "@/maps/schema";
-
-export interface OpenStreetMap {
-    type: string;
-    geometry: OpenStreetMapGeometry;
-    properties: OpenStreetMapProperties;
-}
-
-interface OpenStreetMapGeometry {
-    type: string;
-    coordinates: LatLngTuple;
-}
-
-interface OpenStreetMapProperties {
-    osm_type: "W" | "R" | "N";
-    osm_id: number;
-    extent?: number[];
-    country?: string;
-    state?: string;
-    osm_key: string;
-    countrycode: string;
-    osm_value: string;
-    name: string;
-    type: string;
-    isHidingZone?: boolean;
-    questions?: Question[];
-}
-
-export const OVERPASS_API = "https://overpass-api.de/api/interpreter";
-export const GEOCODER_API = "https://photon.komoot.io/api/";
-
-export const iconColors = {
-    black: "#3D3D3D",
-    blue: "#2A81CB",
-    gold: "#FFD326",
-    green: "#2AAD27",
-    grey: "#7B7B7B",
-    orange: "#CB8427",
-    red: "#CB2B3E",
-    violet: "#9C2BCB",
-};
+    QuestionSpecificLocation,
+} from "./types";
+import { CacheType } from "./types";
 
 export const getOverpassData = async (
     query: string,
@@ -82,42 +41,20 @@ export const determineGeoJSON = async (
         R: "relation",
         N: "node",
     };
-
     const osmType = osmTypeMap[osmTypeLetter];
-
     const query = `[out:json];${osmType}(${osmId});out geom;`;
     const data = await getOverpassData(
         query,
         "Loading map data...",
-        CacheType.PERMANENT_CACHE, // Speedy switching
+        CacheType.PERMANENT_CACHE,
     );
-
     const geo = osmtogeojson(data);
-
     return {
         ...geo,
-        features: geo.features.filter((feature: any) => {
-            if (feature.geometry.type === "Point") {
-                return false;
-            }
-            return true;
-        }),
+        features: geo.features.filter(
+            (feature: any) => feature.geometry.type !== "Point",
+        ),
     };
-};
-
-export const locationFirstTag: {
-    [key in APILocations]: "amenity" | "tourism" | "leisure" | "diplomatic";
-} = {
-    aquarium: "tourism",
-    hospital: "amenity",
-    museum: "tourism",
-    theme_park: "tourism",
-    zoo: "tourism",
-    cinema: "amenity",
-    library: "amenity",
-    golf_course: "leisure",
-    consulate: "diplomatic",
-    park: "leisure",
 };
 
 export const findTentacleLocations = async (
@@ -126,9 +63,7 @@ export const findTentacleLocations = async (
 ) => {
     const query = `
 [out:json][timeout:25];
-nwr["${locationFirstTag[question.locationType]}"="${
-        question.locationType
-    }"](around:${turf.convertLength(
+nwr["${LOCATION_FIRST_TAG[question.locationType]}"="${question.locationType}"](around:${turf.convertLength(
         question.radius,
         question.unit,
         "meters",
@@ -136,50 +71,35 @@ nwr["${locationFirstTag[question.locationType]}"="${
 out center;
     `;
     const data = await getOverpassData(query, text);
-
     const elements = data.elements;
-
     const response = turf.points([]);
-
     elements.forEach((element: any) => {
         if (!element.tags["name"] && !element.tags["name:en"]) return;
-
         if (element.lat && element.lon) {
             const name = element.tags["name:en"] ?? element.tags["name"];
-
             if (
                 response.features.find(
                     (feature: any) => feature.properties.name === name,
                 )
             )
                 return;
-
             response.features.push(
-                turf.point([element.lon, element.lat], {
-                    name,
-                }),
+                turf.point([element.lon, element.lat], { name }),
             );
         }
-
         if (!element.center || !element.center.lon || !element.center.lat)
             return;
-
         const name = element.tags["name:en"] ?? element.tags["name"];
-
         if (
             response.features.find(
                 (feature: any) => feature.properties.name === name,
             )
         )
             return;
-
         response.features.push(
-            turf.point([element.center.lon, element.center.lat], {
-                name,
-            }),
+            turf.point([element.center.lon, element.center.lat], { name }),
         );
     });
-
     return response;
 };
 
@@ -194,58 +114,9 @@ is_in(${latitude}, ${longitude})->.a;
 rel(pivot.a)["admin_level"="${adminLevel}"];
 out geom;
     `;
-
     const data = await getOverpassData(query, "Determining matching zone...");
-
     const geo = osmtogeojson(data);
-
     return geo.features?.[0];
-};
-
-export const geocode = async (address: string, language: string) => {
-    const features = (
-        await (
-            await fetch(`${GEOCODER_API}?lang=${language}&q=${address}`)
-        ).json()
-    ).features as OpenStreetMap[];
-
-    features.forEach((feature) => {
-        feature.geometry.coordinates = convertToLatLong(
-            feature.geometry.coordinates as number[],
-        );
-        if (!feature.properties.extent) return;
-        feature.properties.extent = [
-            feature.properties.extent[1],
-            feature.properties.extent[0],
-            feature.properties.extent[3],
-            feature.properties.extent[2],
-        ];
-    });
-
-    return _.uniqBy(
-        features.filter((feature) => {
-            return feature.properties.osm_type === "R";
-        }),
-        (feature) => feature.properties.osm_id,
-    );
-};
-
-export const determineName = (feature: OpenStreetMap) => {
-    if (feature.properties.state) {
-        return `${feature.properties.name}, ${feature.properties.state}, ${feature.properties.country}`;
-    } else if (feature.properties.country) {
-        return `${feature.properties.name}, ${feature.properties.country}`;
-    } else {
-        return feature.properties.name;
-    }
-};
-
-export const convertToLongLat = (coordinates: LatLngTuple): number[] => {
-    return [coordinates[1], coordinates[0]];
-};
-
-export const convertToLatLong = (coordinates: number[]): LatLngTuple => {
-    return [coordinates[1], coordinates[0]];
 };
 
 export const fetchCoastline = async () => {
@@ -260,16 +131,13 @@ export const fetchCoastline = async () => {
 
 export const trainLineNodeFinder = async (node: string): Promise<number[]> => {
     const nodeId = node.split("/")[1];
-
     const tagQuery = `
 [out:json];
 node(${nodeId});
 wr(bn);
 out tags;
 `;
-
     const tagData = await getOverpassData(tagQuery, "Finding train line...");
-
     const query = `
 [out:json];
 (
@@ -281,52 +149,34 @@ ${tagData.elements
             !element.tags.network
         )
             return "";
-
         let query = "";
-
-        if (element.tags.name) {
-            query += `wr["name"="${element.tags.name}"];`;
-        }
-
-        if (element.tags["name:en"]) {
+        if (element.tags.name) query += `wr["name"="${element.tags.name}"];`;
+        if (element.tags["name:en"])
             query += `wr["name:en"="${element.tags["name:en"]}"];`;
-        }
-
-        if (element.tags["network"]) {
+        if (element.tags["network"])
             query += `wr["network"="${element.tags["network"]}"];`;
-        }
-
         return query;
     })
     .join("\n")}
 );
 out geom;
 `;
-
     const data = await getOverpassData(query, "Finding train lines...");
-
     const geoJSON = osmtogeojson(data);
-
     const nodes: number[] = [];
-
     geoJSON.features.forEach((feature: any) => {
-        // For relations
         if (feature && feature.id && feature.id.startsWith("node")) {
             nodes.push(parseInt(feature.id.split("/")[1]));
         }
     });
-
     data.elements.forEach((element: any) => {
-        // For ways
         if (element && element.type === "node") {
             nodes.push(element.id);
         } else if (element && element.type === "way") {
             nodes.push(...element.nodes);
         }
     });
-
     const uniqNodes = _.uniq(nodes);
-
     return uniqNodes;
 };
 
@@ -347,9 +197,7 @@ export const findPlacesInZone = async (
     timeoutDuration: number = 0,
 ) => {
     let query = "";
-
     const $polyGeoJSON = polyGeoJSON.get();
-
     if ($polyGeoJSON) {
         query = `
 [out:json]${timeoutDuration != 0 ? `[timeout:${timeoutDuration}]` : ""};
@@ -380,21 +228,17 @@ out ${outType};
 `;
     } else {
         const primaryLocation = mapGeoLocation.get();
-
         const additionalLocations = additionalMapGeoLocations
             .get()
             .filter((entry) => entry.added)
             .map((entry) => entry.location);
-
         const allLocations = [primaryLocation, ...additionalLocations];
-
         const relationToAreaBlocks = allLocations
             .map((loc, idx) => {
                 const regionVar = `.region${idx}`;
                 return `relation(${loc.properties.osm_id});map_to_area->${regionVar};`;
             })
             .join("\n");
-
         const searchBlocks = allLocations
             .map((_, idx) => {
                 const regionVar = `area.region${idx}`;
@@ -406,14 +250,12 @@ out ${outType};
                               )
                               .join("\n")
                         : "";
-
                 return `
             ${searchType}${filter}(${regionVar});
             ${altQueries}
           `;
             })
             .join("\n");
-
         query = `
         [out:json]${timeoutDuration !== 0 ? `[timeout:${timeoutDuration}]` : ""};
         ${relationToAreaBlocks}
@@ -423,18 +265,15 @@ out ${outType};
         out ${outType};
         `;
     }
-
     const data = await getOverpassData(
         query,
         loadingText,
         CacheType.ZONE_CACHE,
     );
-
     const subtractedEntries = additionalMapGeoLocations
         .get()
         .filter((e) => !e.added);
     const subtractedPolygons = subtractedEntries.map((entry) => entry.location);
-
     if (subtractedPolygons.length > 0 && data && data.elements) {
         const turfPolys = await Promise.all(
             subtractedPolygons.map(
@@ -447,37 +286,33 @@ out ${outType};
                     ).features[0],
             ),
         );
-
         data.elements = data.elements.filter((el: any) => {
             const lon = el.center ? el.center.lon : el.lon;
             const lat = el.center ? el.center.lat : el.lat;
             if (typeof lon !== "number" || typeof lat !== "number")
                 return false;
             const pt = turf.point([lon, lat]);
-
             return !turfPolys.some((poly) =>
                 turf.booleanPointInPolygon(pt, poly as any),
             );
         });
     }
-
     return data;
 };
 
 export const findPlacesSpecificInZone = async (
-    location: QuestionSpecificLocation,
+    location: `${QuestionSpecificLocation}`,
 ) => {
     const locations = (
         await findPlacesInZone(
             location,
             `Finding ${
-                location === QuestionSpecificLocation.McDonalds
+                location === '["brand:wikidata"="Q38076"]'
                     ? "McDonald's"
                     : "7-Elevens"
             }...`,
         )
     ).elements;
-
     return turf.featureCollection(
         locations.map((x: any) =>
             turf.point([
@@ -488,113 +323,11 @@ export const findPlacesSpecificInZone = async (
     );
 };
 
-export enum QuestionSpecificLocation {
-    McDonalds = '["brand:wikidata"="Q38076"]',
-    Seven11 = '["brand:wikidata"="Q259340"]',
-}
-
-export enum CacheType {
-    CACHE = "jlhs-map-generator-cache",
-    ZONE_CACHE = "jlhs-map-generator-zone-cache",
-    PERMANENT_CACHE = "jlhs-map-generator-permanent-cache",
-}
-
-const determineQuestionCache = _.memoize(() => caches.open(CacheType.CACHE));
-const determineZoneCache = _.memoize(() => caches.open(CacheType.ZONE_CACHE));
-const determinePermanentCache = _.memoize(() =>
-    caches.open(CacheType.PERMANENT_CACHE),
-);
-
-const determineCache = async (cacheType: CacheType) => {
-    switch (cacheType) {
-        case CacheType.CACHE:
-            return await determineQuestionCache();
-        case CacheType.ZONE_CACHE:
-            return await determineZoneCache();
-        case CacheType.PERMANENT_CACHE:
-            return await determinePermanentCache();
-    }
-};
-
-const cacheFetch = async (
-    url: string,
-    loadingText?: string,
-    cacheType: CacheType = CacheType.CACHE,
-) => {
-    try {
-        const cache = await determineCache(cacheType);
-
-        const cachedResponse = await cache.match(url);
-        if (cachedResponse) return cachedResponse;
-
-        if (loadingText) {
-            return toast.promise(
-                async () => {
-                    const response = await fetch(url);
-                    await cache.put(url, response.clone());
-                    return response;
-                },
-                {
-                    pending: loadingText,
-                },
-            );
-        }
-
-        const response = await fetch(url);
-        await cache.put(url, response.clone());
-        return response;
-    } catch (e) {
-        console.log(e); // Probably a caches not supported error
-
-        return fetch(url);
-    }
-};
-
-export const clearCache = async (cacheType: CacheType = CacheType.CACHE) => {
-    try {
-        const cache = await determineCache(cacheType);
-        await cache.keys().then((keys) => {
-            keys.forEach((key) => {
-                cache.delete(key);
-            });
-        });
-    } catch (e) {
-        console.log(e); // Probably a caches not supported error
-    }
-};
-
-export const prettifyLocation = (location: APILocations) => {
-    switch (location) {
-        case "aquarium":
-            return "Aquarium";
-        case "hospital":
-            return "Hospital";
-        case "museum":
-            return "Museum";
-        case "theme_park":
-            return "Theme Park";
-        case "zoo":
-            return "Zoo";
-        case "cinema":
-            return "Cinema";
-        case "library":
-            return "Library";
-        case "golf_course":
-            return "Golf Course";
-        case "consulate":
-            return "Foreign Consulate";
-        case "park":
-            return "Park";
-    }
-};
-
 export const nearestToQuestion = async (
     question: HomeGameMatchingQuestions | HomeGameMeasuringQuestions,
 ) => {
     let radius = 30;
-
     let instances: any = { features: [] };
-
     while (instances.features.length === 0) {
         instances = await findTentacleLocations(
             {
@@ -612,8 +345,6 @@ export const nearestToQuestion = async (
         );
         radius += 30;
     }
-
     const questionPoint = turf.point([question.lng, question.lat]);
-
     return turf.nearestPoint(questionPoint, instances as any);
 };
